// src/app/shared/interfaces/usuario.interface.ts
export interface LoginInterfase {
  email: string;
  password: string;
  remember?: boolean;
}

export interface ListaUsuarios {
  total: number;
  usuarios: UsuarioData[];
}

// Generated by https://quicktype.io

export interface Usuario {
  usuario: UsuarioData;
  token: string;
}

export interface UsuarioData {
  remember?: boolean;
  name: string;
  email: string;
  role: string;
  state: boolean;
  google?: boolean;
  img?: string;
  uid: string;
}

export interface UsuarioLogueado {
  name: string;
  email: string;
  role: string;
  img?: string;
}